Added Share-function + button, background image and some styling
