+++ package.json
@@ -35,4 +35,5 @@
     "babel-preset-expo": "^8.3.0"
   },
   "private": true
+  
 }
